requirements.txt
# Core dependencies
hydra-core>=1.3.0
omegaconf>=2.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning (for future Transformer implementation)
torch>=2.0.0
transformers>=4.30.0

# Development and linting
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pytest>=7.0.0

# Jupyter for analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 